# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevLog+",
    "description": "Single-user developer journal for technical learning and skill maintenance.\n\nDevLog+ has two engines:\n\n- **Learning engine** — builds a Knowledge Profile from journal entries, quizzes, and feedback. Understands what you know, where you're weak, and what to explore next.\n- **Practice engine** — generates weekly Go micro-projects calibrated to your practical level.\n\n## Core workflow\n\n1. **Journal** — write technical entries (typed or dictated) about what you learn.\n2. **Profile** — the system builds and maintains a Knowledge Profile nightly.\n3. **Quiz** — weekly free-text quizzes probe your understanding.\n4. **Readings** — curated reading recommendations from your allowlisted domains.\n5. **Projects** — weekly Go micro-projects with bugs, features, and refactors.\n6. **Feedback** — thumbs-up/down reactions and feedforward notes steer future content.\n7. **Triage** — items the system can't resolve are surfaced for your clarification.\n",
    "license": {
      "name": "Private — single-user local application"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/triage": {
      "get": {
        "tags": [
          "triage"
        ],
        "summary": "List triage items",
        "description": "Return triage items with optional status and severity filters.\n\nTriage items are created automatically by the profile-update,\nquiz-evaluation, and project-evaluation pipelines when the system\nencounters ambiguity it cannot resolve on its own.",
        "operationId": "list_triage_items_api_v1_triage_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TriageStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (pending, accepted, rejected, edited, deferred)",
              "title": "Status"
            },
            "description": "Filter by status (pending, accepted, rejected, edited, deferred)"
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TriageSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity (low, medium, high, critical)",
              "title": "Severity"
            },
            "description": "Filter by severity (low, medium, high, critical)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum items to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum items to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of triage items matching the optional filters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TriageItemResponse"
                  },
                  "title": "Response List Triage Items Api V1 Triage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/triage/{item_id}": {
      "get": {
        "tags": [
          "triage"
        ],
        "summary": "Get triage item",
        "description": "Retrieve a specific triage item by ID.",
        "operationId": "get_triage_item_api_v1_triage__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single triage item with full details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageItemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Triage item not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/triage/{item_id}/resolve": {
      "post": {
        "tags": [
          "triage"
        ],
        "summary": "Resolve triage item",
        "description": "Resolve a triage item with a user action and optional clarification.\n\nAllowed actions: `accepted`, `rejected`, `edited`, `deferred`.  When the\naction is `edited`, `resolution_text` should contain the user's\nclarifying input.",
        "operationId": "resolve_triage_item_api_v1_triage__item_id__resolve_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The resolved triage item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageItemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Triage item not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TriageItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Triage item ID"
          },
          "source": {
            "$ref": "#/components/schemas/TriageSource",
            "description": "Pipeline that created this item (profile_update, quiz_evaluation, project_evaluation)"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id",
            "description": "ID of the source entity (e.g. topic, quiz question) that triggered this item"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short title summarising the issue"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Detailed description of what needs resolution"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Additional context from the pipeline (JSONB)"
          },
          "severity": {
            "$ref": "#/components/schemas/TriageSeverity",
            "description": "Severity: low, medium, high, or critical. High/critical block the next cycle."
          },
          "status": {
            "$ref": "#/components/schemas/TriageStatus",
            "description": "Current status: pending, accepted, rejected, edited, or deferred"
          },
          "resolution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Text",
            "description": "User's clarification text (populated after resolution)"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At",
            "description": "When the item was resolved (null if still pending)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the item was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last modification timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "source_id",
          "title",
          "description",
          "context",
          "severity",
          "status",
          "resolution_text",
          "resolved_at",
          "created_at",
          "updated_at"
        ],
        "title": "TriageItemResponse",
        "description": "An item requiring user attention before the system can proceed."
      },
      "TriageResolveRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TriageStatus",
            "description": "Resolution action: accepted, rejected, edited, or deferred"
          },
          "resolution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Text",
            "description": "Clarifying text from the user. Required when action is 'edited'; optional for other actions.",
            "examples": [
              "The topic should be 'Go error handling' not 'Go exceptions'."
            ]
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "TriageResolveRequest",
        "description": "Resolve a triage item with a user action and optional clarification text."
      },
      "TriageSeverity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "TriageSeverity"
      },
      "TriageSource": {
        "type": "string",
        "enum": [
          "profile_update",
          "quiz_evaluation",
          "project_evaluation"
        ],
        "title": "TriageSource"
      },
      "TriageStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "rejected",
          "edited",
          "deferred"
        ],
        "title": "TriageStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}